fix: paths after structure refactor
