+++ packages/app/src/ui/pages/loading/loading.component.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import LoadingScreen from '../../../submodules/extension/ui/components/ui/loading-screen';
+import LoadingScreen from '../../../../submodules/extension/ui/components/ui/loading-screen';
 
 const Loading = () => {
   return <LoadingScreen showLoadingSpinner />;
